Only hide toolbar in debug log
Split view controller is shi**y enough not to handle navigation
bar properly on rotation.

When in compact mode, the navigation bar comes from the split
view controller. Hiding it and then going to regular (landscape),
will leave the master without the navigation bar and no way to
restore it.