Track module preferences history in Core Data
Restore CDModulePreferencesV3 to track the history of module
prefrences.

This way, excluded endpoints may be saved globally to Core Data
as a starting point. Then in Profile.userInfo we only save the
relevant exclusions for the current configuraiton.

The .excludedEndpoints relationship is therefore moved out of
CDProviderPreferencesV3.

ModuleViewParameters now includes a ModulePreferences observable
that module views can observe.