Use matrix for public_beta and app_review
A tolerant way to cope with scattered approvals. That is, if a
platform build fails to upload, it will not prevent other
platforms from being sent to public_beta/app_review.

The app_store environment is also allowed despite errors, as the
platform builds may have been approved at different times.

This somehow deals with the lottery of getting an approval for
multiple platforms at the same time.

Fixes #1043