Add DNS routes in OpenVPN split setups (#1207)
This was done in the library, but [it was breaking WireGuard
operation](https://github.com/passepartoutvpn/passepartout/issues/1206).
The official WireGuard app doesn't route DNS through the VPN unless the
server is explicitly added to AllowedIPs.

Therefore, move the explicit DNS routes to the OpenVPN implementation
(like it was done in TunnelKit), and maybe remove the code later because
such routes are probably redundant.

Original code from TunnelKit:
https://github.com/passepartoutvpn/tunnelkit/commit/f799f47c256027130e348095b15fca10c8d2387b

Fixes #1206