OpenVPN: Do not route DNS implicitly through the VPN (#1213)
OpenVPN adds routes to the DNS servers when the VPN is not the default
gateway (split setup). This is [an ancient drag from
TunnelKit](https://github.com/passepartoutvpn/tunnelkit/commit/f799f47c256027130e348095b15fca10c8d2387b)
that looks redundant for private DNS inside the VPN (the server already
pushes the routes to reach the DNS server network) and breaking for
public DNS (the VPN may have no Internet access).

There may or may not have been a reason to do that (2019), but it's time
to reconsider the decision. Perhaps this was done in TunnelKit when
routing was not working properly yet, and this code became redundant
later.

Test in split setups before merging.

Final considerations:

- If it's the server pushing the DNS servers, be they public or private,
the server is responsible for making them reachable from the client.
- If it's the client overriding the DNS settings via a module, an option
should be offered to decide whether to route DNS inside or outside of
the VPN (in case a VPN module is present in the profile).