macOS: Fix missing icon in menu bar (#1243)
The first window was closed on launch, but the first window may be the
status bar item. This is always the case when the app is launched as a
login item because there is no main window. Match main window by
internal class name.

"Show" doesn't work on login for the same reason: there's no main
window. Nevertheless, observing that the main window is created when the
app is opened manually, the trick of relaunching "self" seems to work.

Fixes #1103