Introduce provider modules (#1255)
Introduce the new ProviderModule type, then:

- Create a new ProviderView module view
- Move the provider-specific UI from OpenVPNView/WireGuardView to
ProviderView
- In ProfileProcessor, migrate existing profiles and providers to the
new format
- Drop the condition where IPModule used to require an active
ConnectionModule in the profile

The last point is that the condition would not match a ProviderModule
even if it resolves into a ConnectionModule. From a design perspective,
it's the Registry that resolves a ProviderModule into a generic Module
via ProviderModuleResolver, but such a module type cannot be predicted
in Profile.tryBuild() to assert if it's a ConnectionModule or not. At
least, not without a registry. Dropping the condition is just easier,
given that an IPModule without a VPN is harmless/ineffective.

Fixes #1231
Fixes #1221