Resolve memory leaks in tunnel (#1331)
The receipt verifier is a long-running Task and must therefore use weak
self. The verifier prevented the PacketTunnelProvider from being
released, as confirmed by the lack of the "Deinit PTP" message when the
tunnel was stopped. This was not apparent because the tunnel process was
killed regardless.

Resolve [some core
leaks](https://github.com/passepartoutvpn/partout-core/pull/538), too.