Add targets/schemes for Developer ID deployment (#1351)
Set up the foundations for building a standalone Mac app with the tunnel
as a System Extension. In order to keep a single target, extensions must
be embedded conditionally with a script. Use a new "ReleaseMac" scheme
for this purpose.

Embed:

- System Extension if "ReleaseMac" + macOS
- App Extension if any other platform/scheme
- Login Item if macOS

Requirements for the System Extension to work:

- App
- Add a new entitlement `com.apple.developer.system-extension.install =
YES`
- Activate the sysex by submitting a request to the [SystemExtensions
framework](https://developer.apple.com/documentation/systemextensions/ossystemextensionrequest/activationrequest(forextensionwithidentifier:queue:))
- Tunnel
- Use a new "System Extension" target for the same tunnel code, "App
Extension" can't be reused
  - Append `-systemextension` to the `packet-tunnel-provider` entry
- Prepend `NEMachServiceName` in Info.plist with a group from
`com.apple.security.application-groups`

For the tunnel to start at all, the app must be also notarized and
installed in the /Applications folder.

Several things will not work out of the box:

- CloudKit
- Keychain sharing
- App Groups

To be fixed progressively.