Drop environments from ExtendedTunnel
Update Core to handle them in Tunnel.

Do not replicate dataCount and lastErrorCode because they are now
per-profile and directly available via environments. Just notify
with objectWillChange.send() when there are changes in Tunnel so
that the observing views can re-read the values from the updated
environments.

Pinned profile in macOS no longer makes sense with multiple
profiles, and so lastProfileId, which is not necessary in iOS/tvOS
where the NE profile keeps .isEnabled even when inactive.