Improve app and tunnel initialization (#1376)
First, refactor Partout to use PartoutLogger[Context]. Then, simplify
the lifecycle of bootstrap entities in a predictable way. Basically,
move the initialization of the app to AppDelegate, and the tunnel
initialization to PacketTunnelProvider. PartoutLogger is configured the
earliest.

To do this:

- Make AppContext (production) non-static with a convenience init(), so
that its lifecycle and input is predictable
- Move TunnelContext code to PacketTunnelProvider, as it's only used
there
- Move CommonLibrary/UILibrary classes' logic to AppDelegate and
PacketTunnelProvider
- Expose the initialization of KeyValueManager to spot app/shared
preferences
- Wrap shared preferences in AppPreferenceValues

These fixes must keep the current .appex functionality, so restore the
"Logs private data" preference.

For .sysex (Developer ID), #1373, #1374, and #1375 need to be fixed as
well.