Persist tunnel preferences locally (#1379)
App Groups cannot be used in System Extensions, so this PR transmits the
app preferences to the tunnel in a different way. Network Extension
allows an options dictionary to be passed to
NEVPNConnection.startVPNTunnel(options:), which is in turn received by
the tunnel in NEPacketTunnelProvider.startTunnel(options:).

App:

- The app reads the preferences (AppPreference keys) from the
KeyValueManager (backed by the app UserDefaults)
- The manager returns the preferences mapped as AppPreferenceValues
- The AppPreferenceValues object is encoded to Data and indirectly
passed to the .startVPNTunnel() options in ExtendedTunnel
- The options are only sent when the tunnel is started manually from the
app

Tunnel:

- The tunnel receives the encoded Data and decodes it to
AppPreferenceValues
- If the preferences are non-nil, it stores them in the local
UserDefaults and uses them
- If the preferences are nil, they are read from the local UserDefaults
(i.e. last stored values)

Preferences involved:

- dnsFallsBack
- lastUsedProfileId (unused by tunnel)
- logsPrivateData
- skipsPurchases

Fixes #1374