Entitle for products based on the original purchase date (#1412)
The [original build
number](https://developer.apple.com/documentation/storekit/apptransaction/originalappversion)
proved to be platform-specific, which makes it impossible to recognize
artificial purchases across multiple platforms based on the build number
of the first purchase. This is necessary to acknowledge some features to
the old pre-freemium iOS purchasers.

Example: if I bought the iOS app at build 1000 (paid), and I got the
tvOS app at build 2000 (freemium), the tvOS receipt will return 2000 as
the purchased build number, not 1000. Therefore, if I want to grant some
features to users who purchased a build number before 2000, those
features will be available on iOS, but not on tvOS where the build
number is 2000.

Use the [original purchase
date](https://developer.apple.com/documentation/storekit/apptransaction/originalpurchasedate)
in the hope that it makes business model eligibility cross-platform. The
author of [this
post](https://lucas.love/til/change-the-business-model-of-an-app-store-app)
confirms that the [official Apple
advice](https://developer.apple.com/documentation/storekit/supporting-business-model-changes-by-using-the-app-transaction)
is wrong.

Related to #1083 and #1156