Extend paywall with individual purchases (#1422)
We want to offer a "Purchase" button in the purchase confirmation alert,
in addition to the "Connect" and "Save anyway" fallback actions. This is
problematic in case the profile requires the purchase of more than 1
product, because StoreKit only allows 1 purchase at a time.

Therefore, split the paywall into two sections:

- Complete packages (if eligible)
- Individual purchases

In the case of individual purchases, let the user purchase them one at a
time. Gray out the purchased ones, and dismiss the paywall after
purchasing all of them.

The PR prepares for the reintroduction of the paywall on tvOS, but this
will be done later (it crashes now). For now, share the paywall view
logic in PaywallCoordinator and split the iOS/macOS (+Main) and tvOS
(+TV) layouts.