Redesign paywall with individual purchases (#1424)
Rework the paywall and its components on all platforms. Interpose
PaywallView after the coordinator, and rename Main/TV
subviews to Scrollable/Fixed layouts. For now, only the TV uses the
fixed layout. Later, the layout could be chosen based on device size,
but more tweaks are needed for that.

This PR intends to improve the experience of multiple individual
purchases started with #1422. The flow is optimized on tvOS in
particular, where a paywall was not shown at all.

Refactored:

- IncludedFeaturesView → DisclosingFeaturesView (list of features
included in a product)
- AllFeaturesView (list of features highlighting required features and
those included in a product)
- Handle multiple flags in FeatureRow

Stop using StoreKit ProductView in the paywall. Only use it in the
donations screen.