Hide platform-specific behavior of SwiftUI Section (#1426)
When a logical List/Form entry provides a description, this is usually
rendered with a section footer on iOS. This is not the case on macOS,
where inline "row footers" are usually seen. Wrap this annoying
difference in a few "container" modifiers so that:

- iOS: container is shallow, each entry is a Section with a
header/footer
- macOS: container is a Section with a header, each entry is a VStack
with a footer

Namely:

- themeContainer(header, footer)
- themeContainerEntry(header, subtitle, isAction)
- themeContainerWithSingleEntry(header, footer, isAction)

The third one is a convenient shortcut for the frequent scenario where a
section is made of a single row. Footer is rendered as section footer on
iOS, and as row footer on macOS.

Make sure to:

- Always wrap container entries in a .themeContainer(), not
.themeSection(), which is always a Section
- If the container has a header, duplicate the header in the first entry
(iOS only)

This brings more consistency to SwiftUI views because these 3 modifiers
were insanely confusing:

- themeSection()
- themeSectionWithSingleRow()
- themeRowWithSubtitle()