Core products may be suggested unnecessarily (#1434)
Collection.contains() may fail to check if non-essential features
include essential features. .isSuperset(of:) is the right way to check
for set inclusion.


https://forums.swift.org/t/passing-a-collection-to-set-contains-leads-to-confusing-results/76393/2

Luckily, the first condition ensures that Core products are not
suggested if already purchased, but they may be suggested to new
customers for features that don't need them in extreme edge cases.