Integrate new paywall behind feature flag (#1490)
Backport the 3.5.0 paywall in a safe way:

- Create new views rather than modifying existing (`NewPaywallModifier`
entry point)
- Add a .newPaywall config flag
- Pick new or old modifier/views based on the config flag with
DynamicPaywallModifier in the AppCoordinator and ProfileCoordinator
entry points. The fallback paywall alert action also differs in
ProfileCoordinator: the new paywall does nothing (.cancel), the old
paywall saves the profile anyway (.save)
- Rename onCancelPaywall to paywallContinuation, i.e. the action to
perform after dismissing the paywall alert
- Drop the old and unused beta restricted alert

The new paywall improves the purchase flow by presenting a simple
"Purchase" button:

- On connect (AppCoordinator)
- On save profile (ProfileCoordinator)

and works on tvOS too.

The reason why this wasn't done before is because the old paywall was
designed around the StoreKit limitation to only purchase one product at
a time. In fact, it dismisses on a successful purchase. In those cases
where multiple products were required to unlock a set of features (e.g.
On-demand + Apple TV), the old paywall was just unfit, so the app ended
up relying on the "lock icons" to purchase features separately.

The new paywall leverages opacity to visually suggest the path to follow
when purchasing more than one product. The purchases are not atomic, but
the overall L&F tries to mimic an atomic funnel.

Other fixes:

- Incorrect verification on tvOS (IAPManager+Verify.swift)

Fixes #1446