Android: Basic app displaying Partout version (#1521)
Add a CMake configuration to build an NDK library from Partout and
OpenSSL. The Swift runtime is bundled as JNI and the NDK library depends
on it. The unnecessary .so will be stripped later. Automate the import
of Partout and OpenSSL to keep in sync with the binary versions.

Simplify and modernize the demo UI with Jetpack Compose to show a
"Hello" message with the content of partout_version() via JNI.