Activate NESocket with ConfigFlag (#1526)
Roll out migration to
[NESocket](https://github.com/passepartoutvpn/passepartout/pull/1522)
and Swift 6 with a new `.neSocket` flag. Pass to the tunnel via
AppPreferences, and always use NESocket if "Modern cryptography" is
enabled.

Aside, fix a bug where the tunnel would not use "Modern cryptography" if
restarted in the background, e.g., via on-demand rules. That's because
it was using `appPreferences` (renamed to `startPreferences`) rather
than `preferences`, which are non-nil only when the tunnel is started
manually.