Add "Alternative verification" preference (#1528)
Constrain preference to ConfigFlag. Keep the increased verification
interval regardless (1h → 6h).

Refactor KeyValueManager extensions to avoid explicit calls to
PreferenceProtocol.key, by using nominal AppPreference and UIPreference
keys with `for*Preference:`. This is shorter but also safer, because the
risk of using `.rawValue` instead of `.key` is real.

Fixes #1527