Update Partout as dynamic library in Android NDK target (#1531)
Partout is now a dynamic library, and some changes are required for the
Android app to keep working.

New steps for Android:

- Perform the CMake build to generate the vendored libraries
- Build fresh libPartout*.so files with SwiftPM + Android SDK
- Store the libraries in the app JNI folder

The JNI libs are now in `cpp` and split into `partout-<SHA1>` and
`swift-<VERSION>` (the Swift runtime) for convenience. Delete
libswiftObservation.so due to a runtime crash that is not worth
resolving, as we don't use Observation. Remember to create the ABI
subfolders in the `lib/partout-*` and `lib/swift-*` folders, otherwise
the .so files are not copied to the APK.